Add StringID call option
Fixes #43
